JBIDE-17481 - initial impl
JBIDE-17481 - initial impl
