+++ as/plugins/org.jboss.ide.eclipse.archives.webtools/META-INF/MANIFEST.MF
@@ -15,7 +15,6 @@ Require-Bundle: org.eclipse.ui;bundle-version="3.7.0",
  org.eclipse.jem.util;bundle-version="2.1.2",
  org.eclipse.wst.common.modulecore;bundle-version="1.2.100",
  org.apache.ant;bundle-version="1.7.1",
- org.jboss.ide.eclipse.as.core,
  org.eclipse.ui.ide;bundle-version="3.7.0",
  org.eclipse.core.filesystem;bundle-version="1.3.100",
  org.eclipse.ui.navigator;bundle-version="3.5.0",
@@ -26,6 +25,7 @@ Require-Bundle: org.eclipse.ui;bundle-version="3.7.0",
  org.eclipse.pde.core;bundle-version="3.7.0",
  org.jboss.ide.eclipse.archives.core;bundle-version="[3.4.0,4.0.0)",
  org.jboss.ide.eclipse.archives.ui;bundle-version="[3.4.0,4.0.0)",
+ org.jboss.ide.eclipse.as.core,
  org.jboss.ide.eclipse.as.ui;bundle-version="[3.0.0,4.0.0)",
  org.jboss.ide.eclipse.as.wtp.core;bundle-version="[3.0.0,4.0.0)",
  org.jboss.tools.foundation.core;bundle-version="1.1.0"
