+++ as/plugins/org.jboss.ide.eclipse.as.core/META-INF/MANIFEST.MF
@@ -30,7 +30,6 @@ Require-Bundle: org.apache.ant;bundle-version="1.7.1",
  org.jboss.ide.eclipse.as.management.core;bundle-version="[3.0.0,4.0.0)",
  org.jboss.ide.eclipse.as.dmr;bundle-version="[3.0,4.0.0)",
  org.jboss.tools.runtime.core;bundle-version="2.0.0",
- org.jboss.tools.as.runtimes.integration;bundle-version="[3.0.0,4.0.0)",
  org.jboss.ide.eclipse.archives.core;bundle-version="3.5.0",
  org.jboss.tools.foundation.core;bundle-version="1.1.0"
 Bundle-ActivationPolicy: lazy
