+++ as/plugins/org.jboss.ide.eclipse.as.wtp.ui/src/org/jboss/ide/eclipse/as/wtp/ui/composites/IJRECompositeListener.java
@@ -1,0 +1,5 @@
+package org.jboss.ide.eclipse.as.wtp.ui.composites;
+
+public interface IJRECompositeListener {
+	public void vmChanged(AbstractJREComposite comp);
+}
