+++ as/plugins/org.jboss.ide.eclipse.as.wtp.ui/src/org/jboss/ide/eclipse/as/wtp/ui/messages.properties
@@ -17,3 +17,13 @@ ServerAlreadyStartedDialog_Desc=Launching an additional server when one is alrea
 ServerAlreadyStartedDialog_Connect=Set the server adapter to 'started', without launching a new server.
 ServerAlreadyStartedDialog_Launch=Launch a new instance anyway
 
+rwf_DefaultJREForExecEnv=Default JRE for {0}
+wf_JRELabel=JRE
+wf_HomeDirLabel=Home Directory
+browse=Browse...
+wf_NameLabel=Name
+rwf_NameInUse=Runtime name already in use
+rwf_noValidJRE=No valid JREs found for execution environment "{0}"
+rwf_nameTextBlank=The name field must not be blank
+rwf_jboss7homeNotValid=The home directory does not point to a valid server installation.
+rwf_homeIncorrectVersionError=This server type expects a version of {0} but the server directory is of version {1}. This combination may cause critical errors.
