+++ as/plugins/org.jboss.tools.as.runtimes.integration/META-INF/MANIFEST.MF
@@ -13,9 +13,16 @@ Require-Bundle: org.eclipse.core.runtime,
  org.jboss.tools.runtime.core;bundle-version="2.1.0",
  org.jboss.tools.stacks.core;bundle-version="1.0.0",
  org.jboss.ide.eclipse.as.wtp.core;bundle-version="[3.0.0,4.0.0)",
- org.jboss.tools.foundation.core;bundle-version="1.1.0"
+ org.jboss.tools.foundation.core;bundle-version="1.1.0",
+ org.eclipse.ui;bundle-version="3.106.0",
+ org.eclipse.wst.server.ui;bundle-version="1.5.0",
+ org.jboss.ide.eclipse.as.wtp.ui;bundle-version="3.0.0",
+ org.jboss.tools.runtime.ui;bundle-version="3.0.0",
+ org.jboss.tools.foundation.ui;bundle-version="1.1.0"
 Bundle-RequiredExecutionEnvironment: JavaSE-1.6
 Bundle-ActivationPolicy: lazy
 Bundle-ClassPath: .
-Export-Package: org.jboss.tools.as.runtimes.integration.util
+Export-Package: org.jboss.tools.as.runtimes.integration.ui.composites,
+ org.jboss.tools.as.runtimes.integration.ui.wizard,
+ org.jboss.tools.as.runtimes.integration.util
 Bundle-Localization: plugin
