+++ as/plugins/org.jboss.tools.as.runtimes.integration/src/org/jboss/tools/as/runtimes/integration/Messages.properties
@@ -9,3 +9,7 @@ JBossRuntimeStartup_The_JBoss_AS_Hypersonic_embedded_database=The JBoss AS Hyper
 JBossRuntimeStartup_The_JBoss_AS_H2_embedded_database=The JBoss AS H2 embedded database
 LoadRemoteRuntimes=Load Remote Runtimes
 CreateDownloadRuntimes=Create Download Runtimes
+rwf_downloadWarning=Your runtime is being downloaded. If you close this wizard, your download will continue, but your runtime will not be configured.
+rwf_DownloadRuntime=Download and install runtime...
+rwf_downloadTooltipLoading=Checking for compatible runtimes to download.
+rwf_downloadTooltipEmpty=No compatible runtimes available for download.
